rename blur to smooth to match monai
